tests/requirements.txt
requests
pysastrawi